onepage-report/scripts/requirements.txt
python-pptx>=0.6.21
pdfplumber>=0.9.0
Pillow>=9.0.0
# OCR (optional)
# pytesseract>=0.3.10
